% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize_comparison_settings.R
\docType{data}
\name{.clincompare_env}
\alias{.clincompare_env}
\title{Package-Level Settings Environment}
\format{
An object of class \code{environment} of length 2.
}
\usage{
.clincompare_env
}
\description{
Internal environment used to store package settings without modifying
global options.
}
\keyword{internal}
