% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_tolerance.R
\name{get_tolerance}
\alias{get_tolerance}
\title{Get Tolerance Level for Comparisons}
\usage{
get_tolerance()
}
\value{
The current tolerance level as a numeric value.
}
\description{
Retrieves the currently set tolerance level for numeric comparisons.
}
\examples{
\dontrun{
  get_tolerance()
}
}
\keyword{internal}
