% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_datasets.R
\name{print.dataset_comparison}
\alias{print.dataset_comparison}
\title{Print Dataset Comparison Results}
\usage{
\method{print}{dataset_comparison}(x, ...)
}
\arguments{
\item{x}{A \code{dataset_comparison} object from \code{\link[=compare_datasets]{compare_datasets()}}.}

\item{...}{Ignored.}
}
\value{
Invisibly returns \code{x}.
}
\description{
Print Dataset Comparison Results
}
