% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transform_variables.R
\name{transform_variables}
\alias{transform_variables}
\title{Transform Variables in a Dataset}
\usage{
transform_variables(df, transformations)
}
\arguments{
\item{df}{A data frame containing the variables to be transformed.}

\item{transformations}{A list of functions for transforming the variables.
The names of the list should correspond to the variable names in the dataset.}
}
\value{
A data frame with transformed variables.
}
\description{
Applies mathematical or logical transformations to specified columns in a data frame
based on a named list of transformation functions.
}
\examples{
\donttest{
  df <- data.frame(var1 = c("1", "2", "3"), var2 = c("a", "b", "c"),
                   stringsAsFactors = FALSE)
  transform_variables(df, list(var1 = as.numeric, var2 = as.factor))
}
}
\keyword{internal}
